Get cards from User

+++ Atelier3/cpe-asi-dev/cpe-card/src/main/java/repository/CardRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pojo.Card;
 
-public interface CardRepository extends JpaRepository<Card, Long> {
+public interface CardRepository extends JpaRepository<Card, Integer> {
 
 	List<Card> findAll();
